Create a new pr2_sushi stack and copy some useful stuff from pr2_poop_scoop.
In the new stack, there's the skeleton of a util package for reusable stuff, and a sushi package for project specific stuff.

main.py is a bit of a mess. I pulled out a lot of stuff to util but there may still be bits we want so I'm leaving the rest in there.